Try to fix frametime issues with TW. Use eukara's icon in x11, matching the windows icon. Use the extra main-menu options even when there's no menufont available.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5255 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/common.h
@@ -746,7 +746,7 @@ typedef size_t hashfunc_t(unsigned char *digest, size_t maxdigestsize, size_t nu
 hashfunc_t SHA1_m;
 //int SHA1_m(char *digest, size_t maxdigestsize, size_t numstrings, const char **strings, size_t *stringlens);
 //#define SHA1(digest,maxdigestsize,string,stringlen) SHA1_m(digest, maxdigestsize, 1, &string, &stringlen)
-int SHA1(unsigned char *digest, int maxdigestsize, const unsigned char *string, size_t stringlen);
+size_t SHA1(unsigned char *digest, size_t maxdigestsize, const unsigned char *string, size_t stringlen);
 size_t HMAC(hashfunc_t *hashfunc, unsigned char *digest, size_t maxdigestsize, const unsigned char *data, size_t datalen, const unsigned char *key, size_t keylen);
 
 int version_number(void);
